==> preprocessing_zeel/3_preprocessing.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')

beijing_aq_plus_met = pd.read_csv('tmp_beijing_aq_plus_met.csv.gz')
beijing_aq_plus_met.head(2)

### Drop pressure because of too much missing data

beijing_aq_plus_met2 = beijing_aq_plus_met.drop(columns=['pressure'])
beijing_aq_plus_met2.head(2)

### Drop five stations with highest amount of missing data

for station in beijing_aq_plus_met2.station_id.unique():
    tmp_df = beijing_aq_plus_met2[beijing_aq_plus_met2.station_id==station]
    print(station, tmp_df['temperature'].dropna().shape[0]/tmp_df['temperature'].shape[0])

#### Missing data Station IDs: 1009, 1013, 1015, 1020, 1021

beijing_aq_plus_met3 = beijing_aq_plus_met2[~beijing_aq_plus_met2.station_id.isin([1009, 1013, 1015, 1020, 1021])]
beijing_aq_plus_met2.station_id.unique().shape, beijing_aq_plus_met3.station_id.unique().shape

beijing_aq_plus_met3.to_csv('tmp_beijing_aq_plus_met_preprocessed.csv.gz', index=None)
