==> preprocessing_zeel/4_imputing_missing_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import KFold
from sklearn.metrics import mean_squared_error
import warnings
from IPython.display import clear_output
from multiprocessing import Pool
from time import time
warnings.filterwarnings('ignore')

beijing_aq_plus_met = pd.read_csv('tmp_beijing_aq_plus_met_preprocessed.csv.gz')
beijing_aq_plus_met.head(2)

beijing_aq_plus_met['time'] = pd.to_datetime(beijing_aq_plus_met['time'])
beijing_aq_plus_met2 = beijing_aq_plus_met.sort_values(by=['time', 'station_id'])
beijing_aq_plus_met2.head(2)

### Asserting number of data entries per station

for station in beijing_aq_plus_met2.station_id.unique():
    tmp_df = beijing_aq_plus_met2[beijing_aq_plus_met2.station_id==station]
    assert tmp_df.shape == (8760, 15), print(tmp_df.shape)

interpr = ['linear', 'slinear', 'quadratic', 'cubic', 'spline', 'akima']

### Functions

def process_continuous(tmp_df):
    splitter = KFold(n_splits=5, random_state=0, shuffle=True)
    for f_i, (train, test) in enumerate(splitter.split(tmp_df.index)):
        train_index = tmp_df.index[train]
        test_index = tmp_df.index[test]
        
        tmp_df['tmp_data'] = tmp_df[feature]
        tmp_df.loc[test_index, 'tmp_data'] = None
        for algo in interpr:
            filled_vals = tmp_df['tmp_data'].interpolate(method=algo, 
                                                         order=3, 
                                                         limit_direction='both').ffill().bfill()
            tmp_df[algo].loc[test_index] = filled_vals.loc[test_index]
    return tmp_df

def run_k_folds():
    data_df = beijing_aq_plus_met2[['time', 'station_id', feature]]
    for algo in interpr:
        data_df[algo] = None
    
    tmp_df_list = []
    for station in data_df.station_id.unique():
        tmp_df_list.append(data_df[data_df.station_id == station])

    workers = Pool()
    out_list = workers.map(process_continuous, tmp_df_list)
    workers.close()

    final_df = pd.concat(out_list)
    check_df = final_df.dropna(subset=[feature])
    print('###\nRMSE\n###')
    for algo in interpr:
        print(algo, mean_squared_error(check_df[feature], check_df[algo], squared=False))
    print("###")
    return final_df

### Process Continuous variables

init = time()
feature = 'PM25_Concentration'
    
out_df = run_k_folds()
print('time taken:', time()-init, 'seconds')
out_df

init = time()
feature = 'temperature'
    
out_df = run_k_folds()
print('time taken:', time()-init, 'seconds')
out_df

init = time()
feature = 'humidity'
    
out_df = run_k_folds()
print('time taken:', time()-init, 'seconds')
out_df

init = time()
feature = 'wind_speed'
    
out_df = run_k_folds()
print('time taken:', time()-init, 'seconds')
out_df

## Imputing data with best methods

station_wise_df = []
best_methods = {'PM25_Concentration': 'akima', 
                'temperature': 'akima',
                'humidity': 'linear', 
                'wind_speed': 'linear',
                'weather': 'nearest',
                'wind_direction': 'nearest'}
for station in beijing_aq_plus_met2.station_id.unique():
    tmp_df = beijing_aq_plus_met2[beijing_aq_plus_met2.station_id == station]
    for var, method in best_methods.items():
        tmp_df[var] = tmp_df[var].interpolate(method).ffill().bfill()
    station_wise_df.append(tmp_df)
    
filled_df = pd.concat(station_wise_df)
filled_df

for var in best_methods.keys():
    assert filled_df[var].shape == filled_df[var].dropna().shape

final_df = filled_df[['time', 'station_id', 'longitude', 'latitude'] + list(best_methods.keys())]
final_df

final_df.to_csv('../data/processed/final_data.csv.gz', index=None)
